pyproject.toml
[project]
name = "generativeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "faiss-cpu>=1.13.2",
    "groq>=0.37.1",
    "huggingface-hub>=1.4.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "pymupdf>=1.26.7",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
]
